* 4.5.x:
  Reuse test skip condition in tearDown()
  Bump PHPUnit version (#7298)

+++ tests/Functional/Driver/PgSQL/ResultTest.php
@@ -38,8 +38,10 @@ class ResultTest extends FunctionalTestCase
     #[Override]
     protected function tearDown(): void
     {
-        $this->connection->executeStatement('DROP TABLE IF EXISTS types_test');
-        $this->connection->executeStatement('DROP TABLE IF EXISTS types_test2');
+        if (TestUtil::isDriverOneOf('pgsql')) {
+            $this->connection->executeStatement('DROP TABLE IF EXISTS types_test');
+            $this->connection->executeStatement('DROP TABLE IF EXISTS types_test2');
+        }
 
         parent::tearDown();
     }
